requirements.txt
Django==5.0
djangorestframework==3.14.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-htmx==1.17.0
django-cors-headers==4.3.0
Pillow==10.4.0
reportlab==4.0.7
openpyxl==3.1.2
python-dateutil==2.8.2
